+++ backend/src/main/java/br/com/petfamily/canilapi/controller/dto/DespesaRequestDTO.java
@@ -3,6 +3,8 @@ package br.com.petfamily.canilapi.controller.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record DespesaRequestDTO(
@@ -11,7 +13,7 @@ public record DespesaRequestDTO(
 
         @NotNull(message = "O valor da despesa é obrigatório.")
         @Positive(message = "O valor da despesa deve ser maior que zero.")
-        Double valor,
+        BigDecimal valor,
         LocalDate data
 ) {
 }
